Lowers babel version to get to export to work
