+++ .babelrc
@@ -1,3 +1,4 @@
 { 
-  "presets": ["es2015"] 
+  "optional": ["runtime"],
+  "stage": 0
 }
