+++ webpack.config.js
@@ -1,25 +1,28 @@
-/* eslint-disable no-var */
 var webpack = require('webpack');
 var path = require('path');
 
 module.exports = {
   entry: [
-    './src/index.js'
+    './src/index'
   ],
   output: {
-    path: path.join(__dirname, 'dist'),
+    library: 'Cartodb',
     libraryTarget: 'umd',
-    filename: 'bundle.js',
+    filename: 'dist/bundle.js',
+  },
+  resolve: {
+    extensions: ['', '.js', '.jsx']
   },
   devtool: 'eval-source-map',
   plugins: [
+    new webpack.HotModuleReplacementPlugin(),
     new webpack.NoErrorsPlugin()
   ],
   module: {
     loaders: [
       {
-        test: /\.js$/,
-        loaders: ['babel-loader'],
+        test: /\.jsx?$/,
+        loaders: ['react-hot', 'babel'],
         include: path.join(__dirname, 'src')
       }
     ]
